Delete MLIR-AIE submodule (#639)

+++ runtime/src/iree-amd-aie/aie_runtime/test/cdo/add_12_i8_using_2d_dma_op_with_padding.mlir
@@ -4,8 +4,8 @@ module {
   aie.device(npu1_1col) {
     %tile_0_1 = aie.tile(0, 1)
     %objFifo_in0_cons_buff_0 = aie.buffer(%tile_0_1) {address = 0 : i32, sym_name = "objFifo_in0_cons_buff_0"} : memref<64x64xi8>
-    %objFifo_in0_cons_prod_lock = aie.lock(%tile_0_1, 0) {init = 1 : i32, sym_name = "objFifo_in0_cons_prod_lock"}
-    %objFifo_in0_cons_cons_lock = aie.lock(%tile_0_1, 1) {init = 0 : i32, sym_name = "objFifo_in0_cons_cons_lock"}
+    %objFifo_in0_cons_prod_lock = aie.lock(%tile_0_1, 0) {init = 1 : i8, sym_name = "objFifo_in0_cons_prod_lock"}
+    %objFifo_in0_cons_cons_lock = aie.lock(%tile_0_1, 1) {init = 0 : i8, sym_name = "objFifo_in0_cons_cons_lock"}
     %memtile_dma_0_1 = aie.memtile_dma(%tile_0_1) {
       %0 = aie.dma_start(S2MM, 0, ^bb1, ^bb2, repeat_count = 1)
     ^bb1:  // 2 preds: ^bb0, ^bb1
